move faqs config file into app/scripts/front-page/
