+++ packages/openneuro-app/src/scripts/faq/faqs-config.js
